% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResumeServer_delete}
\alias{MouldResumeServer_delete}
\title{按模具编号和出库日期删除}
\usage{
MouldResumeServer_delete(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
按模具编号和出库日期删除
}
\examples{
MouldResumeServer_delete()
}
