% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResumeServer_get}
\alias{MouldResumeServer_get}
\title{同步erp数据到中台}
\usage{
MouldResumeServer_get(input, output, session, dms_token, erp_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
同步erp数据到中台
}
\examples{
MouldResumeServer_get()
}
