% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResumeServer_mould}
\alias{MouldResumeServer_mould}
\title{模具信息查询}
\usage{
MouldResumeServer_mould(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
模具信息查询
}
\examples{
MouldResumeServer_mould()
}
