% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResumeServer_view}
\alias{MouldResumeServer_view}
\title{按模具编号查询模具履历}
\usage{
MouldResumeServer_view(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
按模具编号查询模具履历
}
\examples{
MouldResumeServer_view()
}
